adjustments for cypress and twin.macro + styled-components to work together
