+++ .babelrc.js
@@ -1,5 +1,10 @@
 module.exports = {
-
   presets: [['next/babel', { 'preset-react': { runtime: 'automatic' } }]],
-  plugins: ['babel-plugin-macros', ['styled-components', { ssr: true, displayName: true, preprocess: false }]],
-}
+  plugins: [
+    'babel-plugin-macros',
+    [
+      'babel-plugin-styled-components',
+      { ssr: true, displayName: true, preprocess: false },
+    ],
+  ],
+};
