+++ babel-plugin-macros.config.js
@@ -2,5 +2,10 @@ module.exports = {
   twin: {
     preset: 'styled-components',
     config: './tailwind.config.js',
+    styled: 'styled-components',
+    autoCssProp: true,
+  },
+  styledComponents: {
+    pure: true,
   },
 };
