+++ lib/analytics.js
@@ -1,4 +1,4 @@
-import { fetchGraphQL } from './utils';
+import { fetchGraphQL } from './graphql';
 
 const HASURA_GET_DATA_IMPORTS = `query FrontendGetDataImports {
   ga_data_imports(order_by: {created_at: desc}) {
