+++ lib/articles.js
@@ -1,4 +1,4 @@
-import { fetchGraphQL } from './utils';
+import { fetchGraphQL } from './graphql';
 
 export const HASURA_LIST_ORG_LOCALES = `query FrontendListOrgLocales {
   organization_locales {
