+++ lib/authors.js
@@ -1,4 +1,4 @@
-import { fetchGraphQL } from './utils';
+import { fetchGraphQL } from './graphql';
 
 const HASURA_LIST_AUTHORS = `query FrontendListAuthors {
     organization_locales {
