+++ lib/document.js
@@ -1,8 +1,7 @@
 import fetch from 'node-fetch';
 import AWS from 'aws-sdk';
-import { fetchGraphQL, slugify } from './utils';
+import { fetchGraphQL, slugify } from './graphql';
 import imageSize from 'image-size';
-// const Buffer = require('buffer').Buffer;
 
 export async function processDocumentContents(
   elements,
