+++ lib/homepage.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import { fetchGraphQL } from './utils';
+import { fetchGraphQL } from './graphql';
 
 const HASURA_LIST_HOMEPAGE_LAYOUT_SCHEMAS = `query FrontendListHomepageLayoutSchemas {
   homepage_layout_schemas {
