+++ lib/newsletters.js
@@ -1,4 +1,4 @@
-import { fetchGraphQL } from './utils';
+import { fetchGraphQL } from './graphql';
 
 export const HASURA_LIST_NEWSLETTERS = `query FrontendListNewsletterEditions {
   newsletter_editions(order_by: {newsletter_published_at: desc}) {
