+++ lib/section.js
@@ -1,4 +1,4 @@
-import { fetchGraphQL } from './utils';
+import { fetchGraphQL } from './graphql';
 import _ from 'lodash';
 
 const HASURA_GET_TAG_BY_ID = `query FrontendGetTagByID($id: Int!) {
