+++ lib/site_metadata.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import { fetchGraphQL } from './utils';
+import { fetchGraphQL } from './graphql';
 
 const HASURA_UPSERT_METADATA = `mutation FrontendUpsertMetadata($published: Boolean, $data: jsonb, $locale_code: String) {
   insert_site_metadatas(objects: {published: $published, site_metadata_translations: {data: {data: $data, locale_code: $locale_code}, on_conflict: {constraint: site_metadata_translations_locale_code_site_metadata_id_key, update_columns: data}}}, on_conflict: {constraint: site_metadatas_organization_id_key, update_columns: published}) {
